+++ test/abc.js
@@ -1,11 +1,13 @@
-var assert   = require('assert'),
-    Blast    = require('../index.js'),
-    blastObj = Blast(false),
-    modifiedProto;
-
-modifiedProto = !!(String.prototype.startsWith && Object.divide);
+var assert   = require('assert');
 
 describe('Blast(false)', function() {
+
+	var Blast    = require('../index.js'),
+	    blastObj = Blast(false),
+	    modifiedProto;
+
+	modifiedProto = !!(String.prototype.startsWith && Object.divide);
+
 	it('should not modify the prototype', function() {
 		assert.equal(false, modifiedProto);
 	});
@@ -17,6 +19,13 @@ describe('Blast(false)', function() {
 });
 
 describe('Blast()', function() {
+
+	var Blast;
+
+	it('should apply changes without throwing an error', function() {
+		Blast = require('../index.js')();
+	});
+
 	it('should modify prototype when no parameter is given', function() {
 		assert.equal(true, !!String.prototype.startsWith);
 	});