+++ test/boolean.js
@@ -1,7 +1,11 @@
 var assert = require('assert'),
-    Blast  = require('../index.js')();
+    Blast;
 
 describe('Boolean', function() {
+
+	before(function() {
+		Blast  = require('../index.js')();
+	});
 
 	describe('#toSource()', function() {
 		it('should return the source code representation of the boolean', function() {