+++ test/error.js
@@ -1,12 +1,19 @@
 var assert = require('assert'),
-    Blast  = require('../index.js')();
+    Blast;
 
 describe('Error', function() {
 
+	before(function() {
+		Blast  = require('../index.js')();
+	});
+
 	describe('#toSource()', function() {
 		it('should return the source code representation of the error', function() {
-			var e = new Error('msg');
-			assert.equal('(new Error("msg", undefined, undefined))', e.toSource());
+			var e = new Error('msg'),
+			    src = e.toSource(),
+			    begin = '(new Error("msg", ';
+
+			assert.equal(begin, e.toSource().slice(0,18));
 		});
 	});