+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Fixed lots of bugs in Function.series and Function.parallel
 * Fixed all `arguments` leaks, greatly improving performance
 * Fixed IE bugs, added Function#name getter for IE
+* Added Selenium/Saucelabs testing
 
 ## 0.1.2 (2014-08-28)